Commit to reduce audio complexity
