+++ src/main/java/uk/gov/hmcts/darts/audio/component/AudioMessageDigest.java
@@ -1,0 +1,6 @@
+package uk.gov.hmcts.darts.audio.component;
+
+public interface AudioMessageDigest {
+
+    java.security.MessageDigest getMessageDigest();
+}
