+++ src/main/java/uk/gov/hmcts/darts/audio/service/AudioService.java
@@ -1,8 +1,6 @@
 package uk.gov.hmcts.darts.audio.service;
 
 import com.azure.core.util.BinaryData;
-import org.springframework.web.multipart.MultipartFile;
-import uk.gov.hmcts.darts.audio.model.AddAudioMetadataRequest;
 import uk.gov.hmcts.darts.audio.model.AudioMetadata;
 import uk.gov.hmcts.darts.common.entity.MediaEntity;
 
@@ -14,12 +12,6 @@ public interface AudioService {
     List<MediaEntity> getAudioMetadata(Integer hearingId, Integer channel);
 
     BinaryData encode(Integer mediaId);
-
-    void addAudio(MultipartFile audioFile, AddAudioMetadataRequest addAudioMetadata);
-
-    void linkAudioToHearingInMetadata(AddAudioMetadataRequest addAudioMetadataRequest, MediaEntity mediaEntityToReplace, MediaEntity savedMedia);
-
-    void linkAudioToHearingByEvent(AddAudioMetadataRequest addAudioMetadataRequest, MediaEntity savedMedia);
 
     void setIsArchived(List<AudioMetadata> audioMetadata, Integer hearingId);
