+++ src/main/java/uk/gov/hmcts/darts/audio/service/AudioUploadService.java
@@ -1,0 +1,13 @@
+package uk.gov.hmcts.darts.audio.service;
+
+import org.springframework.web.multipart.MultipartFile;
+import uk.gov.hmcts.darts.audio.model.AddAudioMetadataRequest;
+import uk.gov.hmcts.darts.common.entity.MediaEntity;
+
+public interface AudioUploadService {
+    void addAudio(MultipartFile audioFileStream, AddAudioMetadataRequest addAudioMetadataRequest);
+
+    void linkAudioToHearingInMetadata(AddAudioMetadataRequest addAudioMetadataRequest, MediaEntity mediaToReplace, MediaEntity newMedia);
+
+    void linkAudioToHearingByEvent(AddAudioMetadataRequest addAudioMetadataRequest, MediaEntity savedMedia);
+}
